requirements.txt
gunicorn 
django
djangorestframework
kombu>=4.5.0,<=4.6.5
python-decouple
celery
redis
pika